Memory tracking and reporting now handled by custom utility classes

+++ src/RocketEngine/render/Text.cpp
@@ -71,7 +71,7 @@ void Text::init(std::string fontName, RK_Shader *shader)
 			texture,
 			Vector2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			Vector2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			(uint32)face->glyph->advance.x
 		};
 		characters.insert(std::pair<GLchar, Character>(c, character));
 	}
